Fix scenario page not found
The BrowserRouter uses URL path to keep the UI in sync, and generate
the scenario link like "http://tangleid.github.io/scenario".

It is not suited for Github Pages, navigate to this link may lead
page not found.

To eliminate this problem, replace it with the HashBrowser will use
the URL anchor to keep the UI in sync, and generate the link like
"http://tangleid.github.io/#/scenario".